Refactor topics APIs to accept new name formats
